moved apps into apps directory
