+++ apps/day_counter/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class DayCounterConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'day_counter'
+    name = 'apps.day_counter'
