+++ apps/day_counter/static/jq-js.js
