+++ apps/day_counter/static/style.css
