+++ apps/day_counter/tests/test_models.py
