+++ apps/day_counter/tests/test_views.py
