+++ apps/day_counter/views.py
