+++ apps/users/views.py
@@ -1,4 +1,4 @@
-from users.models import User
+from apps.users.models import User
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.urls import reverse_lazy
 from django.views.generic import UpdateView
