+++ grodvidar/settings/base.py
@@ -34,8 +34,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'users.apps.UsersConfig',
-    'day_counter.apps.DayCounterConfig',
+    'apps.day_counter',
+    'apps.users',
     'crispy_forms',
     'crispy_bootstrap5',
     'allauth',
