+++ grodvidar/urls.py
@@ -20,9 +20,9 @@ from django.urls import path, include
 from grodvidar import settings
 
 urlpatterns = [
-    path('', include('day_counter.urls')),
+    path('', include('apps.day_counter.urls')),
     path('admin/', admin.site.urls),
-    path('users/', include('users.urls')),
+    path('users/', include('apps.users.urls')),
 ]
 if settings.DEBUG:
     urlpatterns += static(settings.MEDIA_URL,
